requirements.txt
pandas==2.2.3
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.1
django-cacheops==7.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3

